Added mod console; cleaner and more modular code
And a lot of improvements in general :)
Open mod console via Project > Show console.
Improvements in the settings system, better error handling.
Improvements to the copyright; changed copyright of the Agent and the
copyright writer.
Copyright writing can now be disabled.
Added comments to various places.
Now doesn't tend to automatically generate mod ID when the checkbox was
not checked in the session before.
*BE SURE TO REPAIR YOUR PROJECTS AS ALWAYS*